backend/requirements.txt
# Core API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-multipart>=0.0.9

# Config & data
pyyaml>=6.0
pandas>=2.2.0
pillow>=10.3.0

# EasyOCR
easyocr>=1.7.1

# Tesseract (system package pytesseract is a wrapper for)
pytesseract>=0.3.13

# HuggingFace
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
accelerate>=0.29.0
sentencepiece>=0.2.0
protobuf>=4.25.0
einops>=0.7.0
timm>=0.9.0

# LMStudio / OpenAI-compat
openai>=1.20.0
httpx>=0.27.0
